requirements.txt
# Web Framework
fastapi
uvicorn
python-multipart

# Deep Learning & Vision
torch
torchvision

# Data & Image Processing
numpy
pandas
Pillow

# OCR
pytesseract

# Utilities
python-dotenv

google-generativeai